Add OpenTelemetryBuilder extensions
Adds extensions for the main `OpenTelemetryBuilder` for registering an OTLP exporter for any combination of traces, metrics and logs. This removes the need to explicitly configure the exporter for individual signal types.

The exporter will only be registered for a signal type if it is enabled and the configured signals to export include the signal type.