Combine IConfiguration and callback
Updates the existing overloads for `AddOtlpExporter` that accept an `IConfiguration` parameter to allow an optional `Action<JaahasOtlpExporterOptions>` to be specified as well, so that settings from configuration can be overridden if required without having to use e.g. `IPostConfigureOptions<T>`.

Also adds `AddOtlpExporter()` and `AddOtlpExporter(string?)` overloads for adding options (named or otherwise) that use the default configuration.