New extensions and defaults (#20)
* Use HttpProtobuf protocol by default

Breaking change. Modifies `JaahasOtlpExporterOptions.Protocol` to specify HTTP/Protobuf export protocol by default instead of gRPC. In most scenarios when exporting to a local collector TLS is not used meaning that gRPC cannot be used.

* Don't hide outer builder references

* Allow config via both IConfiguration and delegate

Breaking change. Modifies the `AddOtlpExporter` extensions for `TracerProviderBuilder`, `MeterProviderBuilder` and `OpenTelemetryLoggerOptions` extensions that accept `IConfiguration` parameters to allow a delegate for performing further configuration to be specified in addition to the configuration object.

* Bump Microsoft.Extensions.Hosting version

* Update version.json

Update to v3.0.0 to reflect breaking changes.